Crate vcl_rust

Rust version of Agner Fog’s vectorclass lib

This crate contains a struct that containts four packed f32 values and uses SIMD instructions to work with them

This crate can only be compiled on x86 or x86 architecture and a proccessor that supports at least sse2 instruction set

Modules

vectorf128

This module contains Vec4f struct with methods and functions to work with it