Struct vcl_rust::vectorf128::Vec4f

pub struct Vec4f { /* private fields */ }

This struct is a wrapper around __m128 intrinsic (or the same intrinsic from x86_64 arch)

Implementations

impl Vec4f

pub const LEN: usize = 4usize

Associated const - size of the packed vector

pub fn new(a: f32, b: f32, c: f32, d: f32) -> Self

Returns Vec4f that contains four f32 values that are equal to the arguments

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(vec, [1.0, 2.0, 3.0, 4.0]);

pub fn from_scalar(value: f32) -> Self

Returns Vec4f that contains four values of type f32 equal to the argument

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::from_scalar(2.0);
assert_eq!(vec, [2.0f32; 4]);

pub fn squared(self) -> Self

Squares every value of the vector and returns a copy

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(vec.squared(), [1.0, 4.0, 9.0, 16.0]);

pub fn pow(self, n: i32) -> Self

Raises every every value of the vector and returns a copy

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(vec.pow(3), [1.0, 8.0, 27.0, 64.0]);

pub fn store(self, buffer: &mut [f32])

Copies values of the vector to a mutable slice

Panics

Panics if buffer.len() is less than 4

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let mut arr = [0.0f32; 4];
vec.store(&mut arr);
assert_eq!(arr, [1.0, 2.0, 3.0, 4.0]);

pub fn store_aligned(self, buffer: &mut [f32])

Copies values of the vector to a mutable slice. Address of the slice must be divisible by 16. Is more efficient than store

Panics

Panics if buffer.len() is less than 4 or the address of the slice is not aligned by 16 bytes

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let mut arr = [0.0f32; 7];
let diff = (arr.as_ptr() as usize % 16) / 4;
let aligned = match diff {
    0 => &mut arr[..4],
    _ => &mut arr[4 - diff..8 - diff],
};
vec.store_aligned(aligned);
assert_eq!(aligned, [1.0, 2.0, 3.0, 4.0]);

pub fn store_aligned_nocache(self, buffer: &mut [f32])

Copies values of the vector to a mutable aligned slice without using cache. Address of the slice must be divisible by 16. Might be more efficient than store_aligned for copying very large arrays of data

Panics

Panics if buffer.len() is less than 4 or the address of the slice is not aligned by 16 bytes

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let mut arr = [0.0f32; 7];
let diff = (arr.as_ptr() as usize % 16) / 4;
let aligned = match diff {
    0 => &mut arr[..4],
    _ => &mut arr[4 - diff..8 - diff],
};
vec.store_aligned_nocache(aligned);
assert_eq!(aligned, [1.0, 2.0, 3.0, 4.0]);

pub fn store_partial(self, buffer: &mut [f32])

Copies values of the vector to a mutable slice. Works for slices with size less than 4

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let mut arr = [0.0f32; 3];
vec.store_partial(&mut arr);
assert_eq!(arr, [1.0, 2.0, 3.0]);

pub fn load(&mut self, buffer: &[f32])

Loads values from float slice

Panics

Panics if buffer.len() is less than 4

Examples

use vcl_rust::vectorf128::Vec4f;

let arr: [f32; 4] = [-2.0, 1.0, 3.0, -4.0];
let mut d = Vec4f::default();
d.load(&arr);
assert_eq!(d, [-2.0, 1.0, 3.0, -4.0]);

pub fn load_aligned(&mut self, buffer: &[f32])

Loads values from float slice aligned by 16 bytes

Panics

Panics if the buffer.len() is less than 4 or if it’s address is not aligned by 16 bytes

Examples

use vcl_rust::vectorf128::Vec4f;

let mut vec = Vec4f::default();
let arr = [2.0f32; 7];
let diff = (arr.as_ptr() as usize % 16) / 4;
let aligned = match diff {
    0 => &arr[..4],
    _ => &arr[4 - diff..8 - diff],
};
vec.load_aligned(aligned);
assert_eq!(aligned, [2.0, 2.0, 2.0, 2.0]);

pub fn load_partial(&mut self, buffer: &[f32])

Copies values from buffer slice to the vector. If buffer.len() is less than 4 fills vector’s tail with zeroes

Examples

use vcl_rust::vectorf128::Vec4f;

let arr = [-2.0, 1.0];
let mut d = Vec4f::default();
d.load_partial(&arr);
assert_eq!(d, [-2.0, 1.0, 0.0, 0.0]);

pub fn insert(self, index: usize, value: f32) -> Self

Inserts f32 value to the chosen index and returns the modified vector

Panics

Panics if index is greater than 3

Examples

use vcl_rust::vectorf128::Vec4f;

let mut a = Vec4f::default();
a = a.insert(0, 6.0);
a = a.insert(3, 23.0);
assert_eq!(a, [6.0, 0.0, 0.0, 23.0]);

pub unsafe fn get_unchecked(&self, index: usize) -> &f32

Returns reference to f32 value by index

Safety

Caller must ensure that index is less than 4

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(unsafe { *vec.get_unchecked(2) }, 3.0);

pub fn get(&self, index: usize) -> Option<&f32>

Return reference to f32 value by index. Returns None if index is greater than 3

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(*vec.get(2).unwrap(), 3.0);

pub fn cutoff(self, size: usize) -> Self

Cuts vector to size, replaces all tail values by zeroes and returns the modified copy

Examples

use vcl_rust::vectorf128::Vec4f;

let e = Vec4f::new(-3.0, 2.0, 1.0, 11.0);
assert_eq!(e.cutoff(2), [-3.0, 2.0, 0.0, 0.0]);

pub fn round(self) -> Self

Rounds all values to closest integer and returns modified copy

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 1.4, 1.5, 1.6);
assert_eq!(vec.round(), [1.0, 1.0, 2.0, 2.0]);

pub fn sqrt(self) -> Self

Returns a vector containing square roots of all values of original vector

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(0.0, 1.0, 2.0, 3.0);
assert_eq!(vec.sqrt(), [0.0, 1.0, f32::sqrt(2.0), f32::sqrt(3.0)]);

pub fn approx_recipr(self) -> Self

Fast approximate calculation of reciprocal (i.e. 1 / self)

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let approx = vec.approx_recipr();
let recipr = Vec4f::from_scalar(1.0) / vec;

let mut arr = [0.0f32; 4];
let diff = approx - recipr;
diff.store(&mut arr);

for value in arr.iter() {
    assert!(value.abs() < 0.01);
}

pub fn approx_rsqrt(self) -> Self

Fast approximate of reverse square root (i.e. 1 / self.sqrt())

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
let approx = vec.approx_rsqrt();
let rsqrt = Vec4f::from_scalar(1.0) / vec.sqrt();

let mut arr = [0.0f32; 4];
let diff = approx - rsqrt;
diff.store(&mut arr);

for value in arr.iter() {
    assert!(value.abs() < 0.01);
}

pub fn change_sign<const I0: bool, const I1: bool, const I2: bool, const I3: bool>( self ) -> Self

Changes signs of vector values chosen by compile-time parameters, returns modified copy

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(-2.0, 3.0, 2.0, -1.0);
assert_eq!(
    vec.change_sign::<true, false, true, false>(),
    [2.0, 3.0, -2.0, -1.0]
);

pub fn abs(self) -> Self

Returns a vector containing absolute values of the original vector

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(-2.0, 3.0, 2.0, -1.0);
assert_eq!(vec.abs(), [2.0, 3.0, 2.0, 1.0]);

pub fn horizontal_add(self) -> f32

Calculates the sum of all vector values

Examples

use vcl_rust::vectorf128::Vec4f;

let vec = Vec4f::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(vec.horizontal_add(), 10.0);